tinker-service/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
tinker-sdk>=0.1.0
